修复: 在视频服务中添加对图像存在性的检查，确保只有在图像存在时才删除文件；更新爬虫设置以使用配置中的超时时间

+++ app/service/video.py
@@ -171,6 +171,7 @@ class VideoService(BaseService):
             exist_path, _ = os.path.split(path)
             for item in ['poster', 'thumb', 'fanart']:
                 image = getattr(exist, item)
-                image_path = os.path.join(exist_path, image)
-                if os.path.exists(image_path):
-                    os.remove(image_path)
+                if image:
+                    image_path = os.path.join(exist_path, image)
+                    if os.path.exists(image_path):
+                        os.remove(image_path)
